Fix cleanup_temp.py again. Windows XP doesn't have don't have LOCALAPPDATA set.
TBR=nsylvain@chromium.org
BUG=
TEST=


Review URL: http://codereview.chromium.org/9653006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@125702 0039d316-1c4b-4281-b951-d872f2087c98

+++ scripts/slave/cleanup_temp.py
@@ -103,13 +103,14 @@ def main_win():
   # Do not add the following cleanup in slaves_utils.py since we don't want to
   # clean them between each test, as the crash dumps may be processed by
   # 'process build' step.
-  crash_reports = os.path.join(
-      os.environ['LOCALAPPDATA'], 'Chromium', 'User Data', 'Crash Reports')
-  if os.path.isdir(crash_reports):
-    for filename in os.listdir(crash_reports):
-      filepath = os.path.join(crash_reports, filename)
-      if os.path.isfile(filepath):
-        os.remove(filepath)
+  if 'LOCALAPPDATA' in os.environ:
+    crash_reports = os.path.join(
+        os.environ['LOCALAPPDATA'], 'Chromium', 'User Data', 'Crash Reports')
+    if os.path.isdir(crash_reports):
+      for filename in os.listdir(crash_reports):
+        filepath = os.path.join(crash_reports, filename)
+        if os.path.isfile(filepath):
+          os.remove(filepath)
   return 0
